move conf stuff to harbor-client (no more anyhow!)
