+++ harbor-client/Cargo.toml
@@ -21,6 +21,7 @@ uuid = { version = "1.8", features = ["v4"] }
 async-trait = "0.1.77"
 bincode = "1.3.3"
 hex = "0.4.3"
+home = "0.5.9"
 
 bitcoin = { version = "0.30.2", features = ["base64"] }
 bip39 = "2.0.0"
