+++ harbor-ui/Cargo.toml
@@ -10,7 +10,6 @@ vendored = ["harbor-client/vendored"]
 [dependencies]
 harbor-client = { version = "0.1.0", path = "../harbor-client" }
 
-anyhow = "1.0.89"
 log = "0.4"
 pretty_env_logger = "0.5" # todo swap to a file logger
 iced = { version = "0.13.1", features = ["debug", "tokio", "svg", "qr_code", "advanced"] }
